feat: 인터뷰 생성 기능 구현

+++ backend/src/main/java/in/backend/core/questionset/application/QuestionSetInfo.java
@@ -6,9 +6,9 @@ import in.backend.core.questionset.entity.QuestionSetRules;
 public record QuestionSetInfo(
         String title,
         Long adminId,
-        Long defaultTailQuestionDepth,
-        Long defaultTimeToThink,
-        Long defaultTimeToAnswer
+        Integer defaultTailQuestionDepth,
+        Integer defaultTimeToThink,
+        Integer defaultTimeToAnswer
 ) {
     public QuestionSetEntity toEntity() {
         var questionSetRules = QuestionSetRules.builder()
